+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Better background handling on Windows
+
 ## [0.1.0] - 2021-08-29
 
 First public version !